don't move these parts into flusher.go and splitter.go
