+++ pkg/stanza/operator/helper/multiline.go
@@ -19,9 +19,101 @@ import (
 	"bytes"
 	"fmt"
 	"regexp"
+	"time"
 
 	"golang.org/x/text/encoding"
 )
+
+// FlusherConfig is a configuration of Flusher helper
+type FlusherConfig struct {
+	Period time.Duration `mapstructure:"force_flush_period"`
+}
+
+// NewFlusherConfig creates a default Flusher config
+func NewFlusherConfig() FlusherConfig {
+	return FlusherConfig{
+		// Empty or `0s` means that we will never force flush
+		Period: time.Millisecond * 500,
+	}
+}
+
+// Build creates Flusher from configuration
+func (c *FlusherConfig) Build() *Flusher {
+	return &Flusher{
+		lastDataChange:     time.Now(),
+		forcePeriod:        c.Period,
+		previousDataLength: 0,
+	}
+}
+
+// Flusher keeps information about flush state
+type Flusher struct {
+	// forcePeriod defines time from last flush which should pass before setting force to true.
+	// Never forces if forcePeriod is set to 0
+	forcePeriod time.Duration
+
+	// lastDataChange tracks date of last data change (including new data and flushes)
+	lastDataChange time.Time
+
+	// previousDataLength:
+	// if previousDataLength = 0 - no new data have been received after flush
+	// if previousDataLength > 0 - there is data which has not been flushed yet and it doesn't changed since lastDataChange
+	previousDataLength int
+}
+
+func (f *Flusher) UpdateDataChangeTime(length int) {
+	// Skip if length is greater than 0 and didn't changed
+	if length > 0 && length == f.previousDataLength {
+		return
+	}
+
+	// update internal properties with new values if data length changed
+	// because it means that data is flowing and being processed
+	f.previousDataLength = length
+	f.lastDataChange = time.Now()
+}
+
+// Flushed reset data length
+func (f *Flusher) Flushed() {
+	f.UpdateDataChangeTime(0)
+}
+
+// ShouldFlush returns true if data should be forcefully flushed
+func (f *Flusher) ShouldFlush() bool {
+	// Returns true if there is f.forcePeriod after f.lastDataChange and data length is greater than 0
+	return f.forcePeriod > 0 && time.Since(f.lastDataChange) > f.forcePeriod && f.previousDataLength > 0
+}
+
+func (f *Flusher) SplitFunc(splitFunc bufio.SplitFunc) bufio.SplitFunc {
+	return func(data []byte, atEOF bool) (advance int, token []byte, err error) {
+		advance, token, err = splitFunc(data, atEOF)
+
+		// Return as it is in case of error
+		if err != nil {
+			return
+		}
+
+		// Return token
+		if token != nil {
+			// Inform flusher that we just flushed
+			f.Flushed()
+			return
+		}
+
+		// If there is no token, force flush eventually
+		if f.ShouldFlush() {
+			// Inform flusher that we just flushed
+			f.Flushed()
+			token = trimWhitespaces(data)
+			advance = len(data)
+			return
+		}
+
+		// Inform flusher that we didn't flushed
+		f.UpdateDataChangeTime(len(data))
+		return
+	}
+}
 
 // Multiline consists of splitFunc and variables needed to perform force flush
 type Multiline struct {
@@ -148,6 +240,24 @@ func NewLineStartSplitFunc(re *regexp.Regexp, flushAtEOF bool) bufio.SplitFunc {
 	}
 }
 
+// SplitNone doesn't split any of the bytes, it reads in all of the bytes and returns it all at once. This is for when the encoding is nop
+func SplitNone(maxLogSize int) bufio.SplitFunc {
+	return func(data []byte, atEOF bool) (advance int, token []byte, err error) {
+		if len(data) >= maxLogSize {
+			return maxLogSize, data[:maxLogSize], nil
+		}
+
+		if !atEOF {
+			return 0, nil, nil
+		}
+
+		if len(data) == 0 {
+			return 0, nil, nil
+		}
+		return len(data), data, nil
+	}
+}
+
 // NewLineEndSplitFunc creates a bufio.SplitFunc that splits an incoming stream into
 // tokens that end with a match to the regex pattern provided
 func NewLineEndSplitFunc(re *regexp.Regexp, flushAtEOF bool) bufio.SplitFunc {
@@ -234,4 +344,79 @@ func trimWhitespaces(data []byte) []byte {
 		return []byte{}
 	}
 	return token
+}
+
+// CustomizedConfig is a configuration of Customized Splitter helper
+type CustomizedConfig struct {
+	Enabled  bool
+	Splitter bufio.SplitFunc
+}
+
+// NewCustomizedConfig  creates a new CustomizedConfig
+func NewCustomizedConfig() CustomizedConfig {
+	return CustomizedConfig{Enabled: false}
+}
+
+// Build creates Customized SplitFunc
+func (c CustomizedConfig) Build(flusher *Flusher) (bufio.SplitFunc, error) {
+	if c.Splitter == nil {
+		return nil, fmt.Errorf("customized splitter is nil")
+	}
+	var splitFunc bufio.SplitFunc
+	if flusher != nil {
+		splitFunc = flusher.SplitFunc(c.Splitter)
+	}
+	return splitFunc, nil
+}
+
+// SplitterConfig consolidates MultilineConfig and FlusherConfig
+type SplitterConfig struct {
+	EncodingConfig EncodingConfig   `mapstructure:",squash,omitempty"`
+	Multiline      MultilineConfig  `mapstructure:"multiline,omitempty"`
+	Flusher        FlusherConfig    `mapstructure:",squash,omitempty"`
+	Customization  CustomizedConfig `mapstructure:",squash,omitempty"`
+}
+
+// NewSplitterConfig returns default SplitterConfig
+func NewSplitterConfig() SplitterConfig {
+	return SplitterConfig{
+		EncodingConfig: NewEncodingConfig(),
+		Multiline:      NewMultilineConfig(),
+		Flusher:        NewFlusherConfig(),
+		Customization:  NewCustomizedConfig(),
+	}
+}
+
+// Build builds Splitter struct
+func (c *SplitterConfig) Build(flushAtEOF bool, maxLogSize int) (*Splitter, error) {
+	enc, err := c.EncodingConfig.Build()
+	if err != nil {
+		return nil, err
+	}
+	if c.Customization.Enabled &&
+		(c.Multiline.LineStartPattern != "" || c.Multiline.LineEndPattern != "") {
+		return nil, fmt.Errorf("multiline should not be set when using customized splitter")
+	}
+	flusher := c.Flusher.Build()
+	var splitFunc bufio.SplitFunc
+	if c.Customization.Enabled {
+		splitFunc, err = c.Customization.Build(flusher)
+	} else {
+		splitFunc, err = c.Multiline.Build(enc.Encoding, flushAtEOF, flusher, maxLogSize)
+	}
+	if err != nil {
+		return nil, err
+	}
+	return &Splitter{
+		Encoding:  enc,
+		Flusher:   flusher,
+		SplitFunc: splitFunc,
+	}, nil
+}
+
+// Splitter consolidates Flusher and dependent splitFunc
+type Splitter struct {
+	Encoding  Encoding
+	SplitFunc bufio.SplitFunc
+	Flusher   *Flusher
 }
